Alternate approach to provider abstractions in the new runtime
When we originally integrated providers into the new runtime, it done
with the assumption that providers and ephemeral resources would be
driven by the execution graph. In practice, we found this too
restrictive and counter intuitive to our goals with ephemeral resources.

After we removed providers and ephemeral resources from the graph, it no
longer made sense to keep most of the provider machinery at that level.
Instead, we can consolidate most of that logic to the unexported
internals of the eval package.

This dramatically simplifies the code around managed providers and
reduces the size of interfaces and integrations between engine and eval.
