+++ internal/engine/applying/apply.go
@@ -8,9 +8,7 @@ package applying
 import (
 	"context"
 
-	"github.com/hashicorp/hcl/v2"
 	"github.com/zclconf/go-cty/cty"
-	"github.com/zclconf/go-cty/cty/function"
 
 	"github.com/opentofu/opentofu/internal/addrs"
 	"github.com/opentofu/opentofu/internal/engine/internal/execgraph"
@@ -61,6 +59,9 @@ func ApplyPlannedChanges(ctx context.Context, plan *plans.Plan, configInst *eval
 	newState, moreDiags := execOps.Finish(ctx)
 	diags = diags.Append(moreDiags)
 
+	moreDiags = oracle.Close(ctx)
+	diags = diags.Append(moreDiags)
+
 	return newState, diags
 }
 
@@ -73,19 +74,4 @@ type evalGlue struct {
 // ResourceInstanceFinalState implements [eval.ApplyGlue].
 func (e *evalGlue) ResourceInstanceFinalState(ctx context.Context, addr addrs.AbsResourceInstance) cty.Value {
 	return e.graph.ResourceInstanceValue(ctx, addr)
-}
-
-// ProviderFunction implements [eval.ApplyGlue].
-func (e *evalGlue) ProviderFunction(ctx context.Context, provider addrs.Provider, providerInstance *addrs.AbsProviderInstanceCorrect, pf addrs.ProviderFunction, rng hcl.Range) (function.Function, tfdiags.Diagnostics) {
-	// Configured
-	if providerInstance != nil {
-		inst, diags := e.ops.providerInstances.ProviderClient(ctx, *providerInstance)
-		if diags.HasErrors() {
-			return function.Function{}, diags
-		}
-		return plugins.BuildProviderFunction(ctx, inst, pf, false, rng)
-	}
-
-	// Unconfigured
-	return e.plugins.BuildFunction(ctx, provider, pf, false, rng)
 }
