+++ internal/engine/applying/operations.go
@@ -12,7 +12,6 @@ import (
 	"slices"
 	"sync"
 
-	"github.com/opentofu/opentofu/internal/engine/internal/common"
 	"github.com/opentofu/opentofu/internal/engine/internal/exec"
 	"github.com/opentofu/opentofu/internal/engine/internal/execgraph"
 	"github.com/opentofu/opentofu/internal/engine/plugins"
@@ -58,8 +57,6 @@ type execOperations struct {
 	// plugins are the provider and provisioner plugins we have available for
 	// use during the apply phase.
 	plugins plugins.Plugins
-
-	providerInstances *common.ProviderInstances
 
 	// Stack of ephemeral and provider close functions
 	// Given the current state of the planning engine, we wait until
@@ -111,7 +108,6 @@ func compileExecutionGraph(ctx context.Context, plan *plans.Plan, oracle *eval.A
 	ops.workingState = plan.PriorState.DeepCopy().SyncWrapper()
 	ops.configOracle = oracle
 	ops.plugins = plugins
-	ops.providerInstances = newProviderInstances(ops)
 
 	return execGraph, compiledGraph, ops, diags
 }
