+++ internal/lang/eval/config_apply.go
@@ -19,6 +19,7 @@ import (
 	"github.com/opentofu/opentofu/internal/lang/eval/internal/configgraph"
 	"github.com/opentofu/opentofu/internal/lang/eval/internal/evalglue"
 	"github.com/opentofu/opentofu/internal/lang/grapheval"
+	"github.com/opentofu/opentofu/internal/providers"
 	"github.com/opentofu/opentofu/internal/tfdiags"
 )
 
@@ -44,16 +45,6 @@ type ApplyGlue interface {
 	// Diagnostics from apply-time actions must be reported through some other
 	// channel controlled by the apply engine itself.
 	ResourceInstanceFinalState(ctx context.Context, addr addrs.AbsResourceInstance) cty.Value
-
-	// Not sure if this should be folded into ResourceInstanceFinalState or not...
-	OpenEphemeralResourceInstance(ctx context.Context, addr addrs.AbsResourceInstance, cfgVal cty.Value, providerInstance addrs.AbsProviderInstanceCorrect) (cty.Value, tfdiags.Diagnostics)
-
-	// ProviderFunction constructs a cty function given a provider and a function address.
-	//
-	// This is a bit odd due to how we support functions on configured providers. We pass in both
-	// a provider address and a provider instance, preferring a call on the configured provider
-	// instance if available.
-	ProviderFunction(ctx context.Context, provider addrs.Provider, providerInstance *addrs.AbsProviderInstanceCorrect, pf addrs.ProviderFunction, rng hcl.Range) (function.Function, tfdiags.Diagnostics)
 }
 
 // ApplyOracle creates an [ApplyOracle] object that can be used to support an
@@ -89,8 +80,25 @@ func (c *ConfigInstance) ApplyOracle(ctx context.Context, glue ApplyGlue) (*Appl
 		return nil, diags
 	}
 
+	managedProviders := newManagedProviders(c.evalContext.Providers, func(ctx context.Context, addr addrs.AbsProviderInstanceCorrect) (cty.Value, tfdiags.Diagnostics) {
+		inst := evalglue.ProviderInstance(ctx, rootModuleInstance, addr)
+		if inst == nil {
+			// We should not get here because the apply phase should only ask for
+			// provider instances that were present during the planning phase, and
+			// we should be using exactly the same configuration source code now.
+			return cty.DynamicVal, tfdiags.New(fmt.Errorf("missing configuration for %s", addr))
+		}
+		v, diags := inst.ConfigValue(ctx)
+		return configgraph.PrepareOutgoingValue(v), diags
+
+	})
+
+	// Inject configured providers
+	evalGlue.providers = managedProviders
+
 	return &ApplyOracle{
-		root: rootModuleInstance,
+		root:      rootModuleInstance,
+		providers: managedProviders,
 	}, diags
 }
 
@@ -99,13 +107,14 @@ func (c *ConfigInstance) ApplyOracle(ctx context.Context, glue ApplyGlue) (*Appl
 // and the specialized API implemented by the apply engine in particular.
 type applyingEvalGlue struct {
 	applyEngineGlue ApplyGlue
+	providers       *managedProviders
 }
 
 // ResourceInstanceValue implements [evalglue.Glue].
 func (g *applyingEvalGlue) ResourceInstanceValue(ctx context.Context, ri *configgraph.ResourceInstance, cfgVal cty.Value, providerInst configgraph.Maybe[*configgraph.ProviderInstance], _ addrs.Set[addrs.AbsResourceInstance]) (cty.Value, tfdiags.Diagnostics) {
 	if ri.Addr.Resource.Resource.Mode == addrs.EphemeralResourceMode {
 		if providerInst, ok := configgraph.GetKnown(providerInst); ok {
-			return g.applyEngineGlue.OpenEphemeralResourceInstance(ctx, ri.Addr, cfgVal, providerInst.Addr)
+			return g.providers.OpenEphemeralResourceInstance(ctx, ri.Addr, cfgVal, ri.Provider, &providerInst.Addr)
 		}
 		log.Printf("[WARN] Provider is not yet known for ephemeral resource %s", ri.Addr)
 		return cty.UnknownVal(cty.DynamicPseudoType), nil
@@ -117,12 +126,11 @@ func (g *applyingEvalGlue) ResourceInstanceValue(ctx context.Context, ri *config
 
 // ProviderFunction implements [evalglue.Glue]
 func (g *applyingEvalGlue) ProviderFunction(ctx context.Context, provider addrs.Provider, providerInst configgraph.Maybe[*configgraph.ProviderInstance], pf addrs.ProviderFunction, rng hcl.Range) (function.Function, tfdiags.Diagnostics) {
-	var providerInstance *addrs.AbsProviderInstanceCorrect
 	if providerInst, ok := configgraph.GetKnown(providerInst); ok {
-		providerInstance = &(providerInst.Addr)
+		return g.providers.ConfiguredFunction(ctx, providerInst.Addr, pf, rng)
 	}
 
-	return g.applyEngineGlue.ProviderFunction(ctx, provider, providerInstance, pf, rng)
+	return g.providers.BuildFunction(ctx, provider, pf, false, rng)
 }
 
 // An ApplyOracle is returned by [ConfigInstance.ApplyOracle] to give the main
@@ -144,7 +152,8 @@ func (g *applyingEvalGlue) ProviderFunction(ctx context.Context, provider addrs.
 // graph that ensures that the apply phase will request information from
 // the oracle only once it has already been made available by earlier work.
 type ApplyOracle struct {
-	root evalglue.CompiledModuleInstance
+	root      evalglue.CompiledModuleInstance
+	providers *managedProviders
 }
 
 // DesiredResourceInstance returns the [DesiredResourceInstance] object
@@ -206,18 +215,8 @@ func (o *ApplyOracle) DesiredResourceInstance(ctx context.Context, addr addrs.Ab
 // to refer only to provider instances that are present ni the configuration.
 // If this _does_ return cty.NilVal then that suggests a bug in the planning
 // engine, causing it to create an incorrect execution graph.
-func (o *ApplyOracle) ProviderInstanceConfig(ctx context.Context, addr addrs.AbsProviderInstanceCorrect) (cty.Value, tfdiags.Diagnostics) {
-	inst := evalglue.ProviderInstance(ctx, o.root, addr)
-	if inst == nil {
-		// We should not get here because the apply phase should only ask for
-		// provider instances that were present during the planning phase, and
-		// we should be using exactly the same configuration source code now.
-		var diags tfdiags.Diagnostics
-		diags = diags.Append(fmt.Errorf("missing configuration for %s", addr))
-		return cty.DynamicVal, diags
-	}
-	v, diags := inst.ConfigValue(ctx)
-	return configgraph.PrepareOutgoingValue(v), diags
+func (o *ApplyOracle) ProviderInstance(ctx context.Context, addr addrs.AbsProviderInstanceCorrect) (providers.Interface, tfdiags.Diagnostics) {
+	return o.providers.ProviderInstance(ctx, addr)
 }
 
 // AnnounceAllGraphevalRequests calls the given function once for each internal
@@ -234,4 +233,8 @@ func (o *ApplyOracle) ProviderInstanceConfig(ctx context.Context, addr addrs.Abs
 // going to use it for something.
 func (o *ApplyOracle) AnnounceAllGraphevalRequests(announce func(workgraph.RequestID, grapheval.RequestInfo)) {
 	o.root.AnnounceAllGraphevalRequests(announce)
+}
+
+func (o *ApplyOracle) Close(ctx context.Context) tfdiags.Diagnostics {
+	return o.providers.Close(ctx)
 }
